build: consolidate tsconfig
I think given that the project is all ESM it seems fine to consolidate
down to one tsconfig. I ran the build with this config and the previous
config and diffed the files outputted to dist and they are identical. No
worries if you don't want to merge after the recent churn/issues. I
looked at the docs for each of the changes I made, there are no issues
reported by  and the tests still pass.